Make `new` overridable.
The documentation for exconstructor suggests that you can override the
new function in modules that use ExConstructor.  This might be the case,
but without marking them with `defoverridable` we can't use `super` to
easily call into the original new function.

The case where I've found this useful is for nested structs - I have a
struct that contains a list of other structs, and I'd like to have my
new call convert all the members of the list as well.