Fix typespec for populate_struct/2
When I did the cleanup in #27 I removed a typespec:
https://github.com/appcues/exconstructor/pull/27/files#diff-9d740b81ce2e444f305074e30bf031039b97d234ac866fa5b329f46902052d6fL182

This was because having multiple typespecs for a given function is
ignored:

```
Overloaded contract for ExConstructor.populate_struct/3 has
overlapping domains; such contracts are currently unsupported and
are simply ignored.
```

Removing that bit, while correct, made the remaining typespec for
populate_struct/3 *wrong*, as the options argument can also be
`map_or_kwlist`, so I changed the type signature to be correct there.